refactor: change component name in playground
