+++ packages/_vue3-migration-test/src/components/index.ts
@@ -3,4 +3,4 @@ export { default as TestBaseColumnPickerDropdown } from './column-picker/test-ba
 export { default as TestBaseColumnPickerList } from './column-picker/test-base-column-picker-list.vue';
 export { default as TestBaseDropdown } from './test-base-dropdown.vue';
 export { default as TestBaseEventButton } from './test-base-event-button.vue';
-export { default as TestLayoutsComposable } from './test-layouts-composable.vue';
+export { default as TestUseLayouts } from './test-use-layouts.vue';
